% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expit}
\alias{expit}
\title{inverse logit function}
\usage{
expit(logp)
}
\arguments{
\item{logp}{a real-valued number}
}
\value{
the inverse logit of logp, which will be between 0 and 1
}
\description{
inverse logit function
}
\author{
Jay Ver Hoef
}
