% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLdd.R
\name{LLdd}
\alias{LLdd}
\title{loglikelihood for model}
\usage{
LLdd(
  N_pup1,
  N_adu1,
  delta,
  phi,
  kappa1,
  rho,
  pdelta_mu,
  pdelta_sd,
  pphi_mu,
  pphi_sd,
  pkappa_mu,
  pkappa_sd,
  prho_mu,
  prho_sd,
  pN_pup1_mu,
  pN_pup1_sd,
  pN_adu1_mu,
  pN_adu1_sd,
  pp,
  H_adu,
  H_pup,
  nyrs
)
}
\arguments{
\item{N_pup1}{population value for pups in first year}

\item{N_adu1}{population value for adults in first year}

\item{delta}{Leslie matrix parameter for adult survival}

\item{phi}{Leslie matrix parameter for adult fecundity (both sexes, until survey time)}

\item{kappa1}{Leslie matrix parameter for pup survival after surveys}

\item{rho}{parameter for density dependent decay in fecundity}

\item{pdelta_mu}{mean of normal distribution for prior for logit(delta)}

\item{pdelta_sd}{standard deviation of normal distribution for prior for logit(delta)}

\item{pphi_mu}{mean of normal distribution for prior for logit(phi)}

\item{pphi_sd}{standard deviation of normal distribution for prior for logit(phi)}

\item{pkappa_mu}{mean of normal distribution for prior for logit(kappa)}

\item{pkappa_sd}{standard deviation of normal distribution for prior for logit(kappa)}

\item{prho_mu}{mean of normal distribution for prior for log(rho)}

\item{prho_sd}{standard deviation of normal distribution for prior for log(rho)}

\item{pN_pup1_mu}{mean of normal distribution for prior for N_pup1}

\item{pN_pup1_sd}{standard devitation of normal distribution for prior for N_pup1}

\item{pN_adu1_mu}{mean of normal distribution for prior for N_adu1}

\item{pN_adu1_sd}{standard devitation of normal distribution for prior for N_adu1}

\item{pp}{data set with pup production}

\item{H_adu}{vector of adult harvest for each year}

\item{H_pup}{vector of pup harvest for each year}

\item{nyrs}{number of years (one more than length of harvest vectors)}
}
\value{
the loglikelihood
}
\description{
evaluates to the loglikelihood for MCMC sampling
}
\author{
Jay Ver Hoef
}
