% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCharp.R
\name{MCMCharp}
\alias{MCMCharp}
\title{MCMC sampling}
\usage{
MCMCharp(
  burnin = 10000,
  niter = 1e+05,
  thin = 100,
  pdelta_mu = 2.2,
  pdelta_sd = 0.5,
  pphi_mu = -0.4,
  pphi_sd = 0.5,
  pkappa_mu = 1.4,
  pkappa_sd = 0.5,
  prho_mu = -4,
  prho_sd = 0.5,
  pN_pup1_mu = 3e+05,
  pN_pup1_sd = 50000,
  pN_adu1_mu = 1e+06,
  pN_adu1_sd = 1e+05,
  N_pup1_tune = 10000,
  N_adu1_tune = 10000,
  delta_tune = 0.05,
  phi_tune = 0.05,
  kappa_tune = 0.05,
  rho_tune = 0.03,
  harvadu_data,
  harvpup_data,
  pupcount_data,
  set_seed = NULL
)
}
\arguments{
\item{burnin}{the number of MCMC samples for burnin. No MCMC samples will be saved until burnin has completed.}

\item{niter}{the number of MCMC iterations after burnin, when samples will be saved.}

\item{thin}{thinning of the MCMC chain.  This is the number of samples to skip over before saving a sample.}

\item{pdelta_mu}{mean of normal distribution for prior for logit(delta), where delta is the Leslie matrix parameter for adult survival. Default is 2.2}

\item{pdelta_sd}{standard deviation of normal distribution for prior for logit(delta). Default is 0.5}

\item{pphi_mu}{mean of normal distribution for prior for logit(phi), where phi is the Leslie matrix parameter for adult adult fecundity (both sexes, until survey time). Default is -0.4}

\item{pphi_sd}{standard deviation of normal distribution for prior for logit(phi). Default is 0.5}

\item{pkappa_mu}{mean of normal distribution for prior for logit(kappa), where kappa is the Leslie matrix parameter for pup survival after surveys. Default is 1.4}

\item{pkappa_sd}{standard deviation of normal distribution for prior for logit(kappa). Default is 0.5}

\item{prho_mu}{mean of normal distribution for prior for log(rho), where rho is a parameter for density dependent decay in fecundity. Default is -4.0}

\item{prho_sd}{standard deviation of normal distribution for prior for log(rho). Default is 0.5}

\item{pN_pup1_mu}{mean of normal distribution for prior for N_pup1, the population value for pups in first year. Default is 300000}

\item{pN_pup1_sd}{standard devitation of normal distribution for prior for N_pup1. Default is 50000}

\item{pN_adu1_mu}{mean of normal distribution for prior for N_adu1, the population value for adults in the first year. Default 1000000}

\item{pN_adu1_sd}{standard devitation of normal distribution for prior for N_adu1. Default is 100000}

\item{N_pup1_tune}{Tuning the Metropolis proposal, which is the standard deviation of a normal distribution centered on the current value for N_pup1. Default is 10000}

\item{N_adu1_tune}{Tuning the Metropolis proposal, which is the standard deviation of a normal distribution centered on the current value for N_adu1. Default is 10000}

\item{delta_tune}{Tuning the Metropolis proposal, which is the standard deviation of a normal distribution centered on the current value for logit(delta). Default is 0.05}

\item{phi_tune}{Tuning the Metropolis proposal, which is the standard deviation of a normal distribution centered on the current value for logit(phi). Default 0.05}

\item{kappa_tune}{Tuning the Metropolis proposal, which is the standard deviation of a normal distribution centered on the current value for logit(kappa). Default is 0.05}

\item{rho_tune}{Tuning the Metropolis proposal, which is the standard deviation of a normal distribution centered on the current value for log(rho). Default is 0.03}

\item{harvadu_data}{vector of adult harvest values for each year}

\item{harvpup_data}{vector of adult harvest values for each year}

\item{pupcount_data}{data set of pup counts}

\item{set_seed}{default If this is specified, the seed value for randomization will be set, so all results will be repeatable.  If NULL, the seed value is taken from the computer clock. Default is NULL.}
}
\value{
a list of the MCMC chain values (first six list items), as well as acceptance rates (last 6 list items).
}
\description{
MCMC sampling for simple harp seal model with a 2-age Leslie matrix model where parameters are stochastic from year to year, but each parameter has a common prior across years.  There are also priors on starting values for adults and pups, and a prior that controls density-dependent decay in fecundity.
}
\examples{
library(MCMCharp) 
data(catch_data)
data(pup_production)
#use all defaults
W = MCMCharp(
  harvadu_data = catch_data$adu, 
  harvpup_data = catch_data$pup,
  pupcount_data = pup_production, 
  set_seed = 1001
)
}
\author{
Jay Ver Hoef
}
